refactor: extract template
